Add files via upload
To what extent do attitudes toward immigration explain the switching of votes of 2012 Obama supporters who became 2016 Trump supporters?

We use data from the 2016 Cooperative Congression Election Study to attempt to answer this. The population of interest is taken to be all people who voted for Barack Obama in the 2012 Presidential general election who were eligible to vote in the 2016 Presidential general election. In the first section, we will see to what extent member of this population switching their votes to Donald Trump can be explained by individual demographic variables: gender, education, race, and party identification. We also construct an index to describe attitudes toward immigration and see how this relates to vote switching. In the second section, we test out models with immigration and one other explanatory variable to predict vote switching, looking for interaction. In all cases, we use the CCES survey weights to improve the representativeness of the data.